Obsolete ReceiveFeature and move logic into ReceiveComponent
