+++ src/NServiceBus.Core/obsoletes.cs
@@ -1112,8 +1112,8 @@ namespace NServiceBus
     // Just to make sure we remove it in v8. We keep it around for now just in case some external feature
     // depended on it using `DependsOn(string featureTypeName)`
     [ObsoleteEx(
-           RemoveInVersion = "8",
-           TreatAsErrorFromVersion = "7")]
+        RemoveInVersion = "8",
+        TreatAsErrorFromVersion = "7")]
     class HostStartupDiagnostics : Feature
     {
         public HostStartupDiagnostics()
@@ -1127,4 +1127,25 @@ namespace NServiceBus
     }
 }
 
+namespace NServiceBus
+{
+    using Features;
+
+    // Just to make sure we remove it in v8. We keep it around for now just in case some external feature
+    // depended on it using `DependsOn(string featureTypeName)`
+    [ObsoleteEx(
+        RemoveInVersion = "8",
+        TreatAsErrorFromVersion = "7")]
+    class ReceiveFeature : Feature
+    {
+        public ReceiveFeature()
+        {
+            EnableByDefault();
+        }
+
+        protected internal override void Setup(FeatureConfigurationContext context)
+        {
+        }
+    }
+}
 #pragma warning restore 1591
